Sync DeliveryManager
Implemented client-authoritative synchronization for DeliveryManager.

NOTE: Only the server is responsible for generating the recipes and broadcasting messages to all clients. The client, on the other hand, validates the plates and informs the server whether they are correct or incorrect. The server then processes that RPC and broadcasts the result to every client.